ref(navbar): Single source of truth for navbar items
